CARDS-1886: Replace Torch import with Clarity import Move generic filters and mappers into the common clarity-integration module
