+++ modules/clarity-integration/src/main/java/io/uhndata/cards/clarity/importer/internal/ConfiguredDiscardFilter.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package io.uhndata.cards.prems.internal.importer;
+package io.uhndata.cards.clarity.importer.internal;
 
 import java.util.Map;
